pyproject.toml
[project]
name = "bell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "lukaspodobnik", email = "lukas.podobnik@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[project.scripts]
bell = "bell:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
